requirements.txt
fastapi==0.115.6
mangum==0.17.0
openai==0.28.0
pydantic==2.10.4
python-dotenv==1.0.1
